Update LandDetails and LandOwner services for user-specific behavior
Refactored `LandDetailsService` and `LandOwnerService` to enhance functionality with user-specific actions, such as fetching and saving land details based on user IDs. Streamlined method signatures and naming for clarity, updated repositories for ID consistency, and improved handling of related entities like `LandOwner` documents.

+++ src/main/java/com/example/aurigraph/farmers/DTO/CompleteLandDetailsOutDTO.java
@@ -149,7 +149,7 @@ public class CompleteLandDetailsOutDTO {
 
     private List<Witness> witnesses;
 
-    private Integer userId;
+    private Long userId;
 
     public boolean isApproved(){
         return approved;
@@ -212,11 +212,11 @@ public class CompleteLandDetailsOutDTO {
 
     private Timestamp lastModifiedDate ;
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
