+++ src/main/java/com/example/aurigraph/farmers/Repository/LandDetailsRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Repository
 public interface LandDetailsRepository extends CrudRepository<LandDetails, Long> {
 
-    List<LandDetails> findLandDetailsByUserId(Integer userId);
+    List<LandDetails> findLandDetailsByUserId(Long userId);
 
 
 }
